requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.1.0
eth_abi==5.1.0
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
six==1.16.0
toolz==0.12.1
typing_extensions==4.11.0
urllib3==2.2.1
web3==6.18.0
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
